requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
mlflow
python-box
fastapi
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
notebook
jupyter
python-dotenv
-e .
